+++ astropy/nddata/tests/test_utils.py
@@ -1,5 +1,6 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 
+from packaging.version import Version
 import pytest
 import numpy as np
 from numpy.testing import assert_allclose, assert_array_equal
@@ -9,7 +10,6 @@ from astropy.nddata.utils import (extract_array, add_array, subpixel_indices,
                                   block_reduce, block_replicate,
                                   overlap_slices, NoOverlapError,
                                   PartialOverlapError, Cutout2D)
-
 from astropy.wcs import WCS, Sip
 from astropy.wcs.utils import proj_plane_pixel_area
 from astropy.coordinates import SkyCoord
@@ -281,7 +281,7 @@ def test_extract_array_return_pos():
     The result will differ by mode. All test here are done in 1d because it's
     easier to construct correct test cases.
     '''
-    large_test_array = np.arange(5)
+    large_test_array = np.arange(5, dtype=float)
     for i in np.arange(-1, 6):
         extracted, new_pos = extract_array(large_test_array, 3, i,
                                            mode='partial',
@@ -297,6 +297,14 @@ def test_extract_array_return_pos():
         extracted, new_pos = extract_array(large_test_array, (3,), (i,),
                                            mode='trim', return_position=True)
         assert new_pos == (expected, )
+
+
+def test_extract_array_nan_fillvalue():
+    if Version(np.__version__) >= Version('1.20'):
+        msg = 'fill_value cannot be set to np.nan if the input array has'
+        with pytest.raises(ValueError, match=msg):
+            extract_array(np.ones((10, 10), dtype=int), (5, 5), (1, 1),
+                          fill_value=np.nan)
 
 
 def test_add_array_odd_shape():